Type casting

+++ src/server/mod.rs
@@ -645,9 +645,9 @@ impl Controller for ControllerService {
                     println!("Trace route: {}", traceroute);
                     // Sleep 10 seconds to give the client time to finish the task and receive the last responses
                     if traceroute {
-                        tokio::time::sleep(Duration::from_secs((45 + clients.len() - client_id) as u64)).await;
+                        tokio::time::sleep(Duration::from_secs(45 + clients.len() as u64 - client_id as u64)).await;
                     } else {
-                        tokio::time::sleep(Duration::from_secs((10 + clients.len() - client_id) as u64)).await;
+                        tokio::time::sleep(Duration::from_secs(10 + clients.len() as u64 - client_id as u64)).await;
                     }
                     println!("[Server] Letting client with ID {} know it the measurement is finished", client_id);
                     // Send a message to the client to let it know it has received everything for the current task
